add navbar menu and social components
