+++ package.json
@@ -25,6 +25,7 @@
     "bootstrap": "^5.3.2",
     "firebase": "^10.7.1",
     "is-number": "^7.0.0",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.1",
     "react-dom": "^18.2.0",
