+++ src/components/Navbar.jsx
@@ -1,28 +1,12 @@
 import styles from "./Navbar.module.css";
-import React, { useState } from "react";
-import { Link, NavLink } from "react-router-dom";
+import { useState } from "react";
+import { Link } from "react-router-dom";
 import logo from "../assets/logo.png";
-
-// import ICONS from "../assets/social_media";
+import MenuItems from "./Navbar/MenuItems";
+import SocialLinks from "./Navbar/SocialLinks";
 
 export const Navbar = () => {
   const [menuOpen, setMenuOpen] = useState(false);
-
-  const menuItems = [
-    // { to: "/about-foresee", label: "About Foresee" },
-    { to: "/", label: "About Foresee" },
-    { to: "/join-foresee", label: "Join Foresee" },
-    { to: "/events", label: "Events" },
-    { to: "/educational-games", label: "Games" },
-  ];
-
-  const icons = [
-    "src/assets/social_media/EmailUs.svg",
-    "src/assets/social_media/Facebook.svg",
-    "src/assets/social_media/Instagram.svg",
-    "src/assets/social_media/LinkedIn.svg",
-    "src/assets/social_media/Youtube.svg",
-  ];
 
   return (
     <nav className={styles.nav}>
@@ -43,21 +27,8 @@ export const Navbar = () => {
         <span className={styles.bar}></span>
       </div>
       <ul className={menuOpen ? styles.open : ""}>
-        {menuItems.map((item, index) => (
-          <li key={index} onClick={() => setMenuOpen(!menuOpen)}>
-            <NavLink
-              className={({ isActive }) => (isActive ? styles.linkActive : "")}
-              to={item.to}
-            >
-              {item.label}
-            </NavLink>
-          </li>
-        ))}
-        <div className={styles.iconsDiv}>
-          {icons.map((src, index) => (
-            <img height={"30px"} key={index} src={src} alt={`Image ${index}`} />
-          ))}
-        </div>
+        <MenuItems menuOpen={menuOpen} setMenuOpen={setMenuOpen} />
+        <SocialLinks menuOpen={menuOpen} />
       </ul>
     </nav>
   );
